+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.0.2 - 2025-03-23
+
+#### Noteworthy Changes
+
+* Adds ability to set the service when creating domains through easyDNS API
+* Adds support for SSHFP, TLSA, DS and ROOT NS records
+* Adds new check to ensure DS records are not created at root
+
 ## v0.0.1 - 2022-01-07 - Moving
 
 #### Nothworthy Changes